Import BioentityPropertyName after moving to a new package 'bioentities'
